update packageName from management to v1
the generator will append the package name to the model and data type
definition:
for example if we have Volume definition we will see v1Volume in case v1
is the package name.
Due toa change we did in the package name to express the management API
from management to package lightbits.lightbox.v1 we needed to change the
string replace as well.